# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: YookPay African Payment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: getTransactions
      tags: [transactions]
      summary: Get transaction history
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: number
          required: false
        - in: query
          name: limit
          schema:
            type: number
          required: false
        - in: query
          name: status
          schema:
            type: string
            enum: [PENDING, SUCCESS, FAILED]
          required: false
        - in: query
          name: currency
          schema:
            type: string
          required: false
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
  /transactions/deposit:
    post:
      operationId: createDeposit
      tags: [transactions]
      summary: Create a deposit via Mobile Money
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositBody"
      responses:
        "201":
          description: Deposit initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/withdraw:
    post:
      operationId: createWithdrawal
      tags: [transactions]
      summary: Create a withdrawal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawBody"
      responses:
        "201":
          description: Withdrawal initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/transfer:
    post:
      operationId: createTransfer
      tags: [transactions]
      summary: Transfer between wallets
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferBody"
      responses:
        "201":
          description: Transfer initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/fee-preview:
    post:
      operationId: getFeePreview
      tags: [fees]
      summary: Preview fees for a transaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeePreviewBody"
      responses:
        "200":
          description: Fee breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeBreakdown"
  /transactions/webhook:
    post:
      operationId: processWebhook
      tags: [transactions]
      summary: Process payment webhook callback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookBody"
      responses:
        "200":
          description: Webhook processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Transaction:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        type:
          type: string
          enum: [DEPOSIT, WITHDRAWAL, TRANSFER]
        status:
          type: string
          enum: [PENDING, SUCCESS, FAILED]
        amount:
          type: number
        fee:
          type: number
        netAmount:
          type: number
        currency:
          type: string
        country:
          type: string
          nullable: true
        operator:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        reference:
          type: string
        feeRate:
          type: number
          nullable: true
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - status
        - amount
        - fee
        - netAmount
        - currency
        - reference
        - createdAt
        - updatedAt
    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
        totalPages:
          type: number
      required:
        - transactions
        - total
        - page
        - limit
        - totalPages
    TransactionResponse:
      type: object
      properties:
        transaction:
          $ref: "#/components/schemas/Transaction"
        feeBreakdown:
          $ref: "#/components/schemas/FeeBreakdown"
        message:
          type: string
      required:
        - transaction
        - feeBreakdown
    DepositBody:
      type: object
      properties:
        amount:
          type: number
          minimum: 100
        country:
          type: string
          enum: [CM, SN, CD]
        operator:
          type: string
          enum: [MTN, ORANGE, MOOV, WAVE]
        phone:
          type: string
      required:
        - amount
        - country
        - operator
        - phone
    WithdrawBody:
      type: object
      properties:
        amount:
          type: number
          minimum: 100
        currency:
          type: string
          enum: [XAF, XOF, CDF]
        phone:
          type: string
        operator:
          type: string
          enum: [MTN, ORANGE, MOOV, WAVE]
      required:
        - amount
        - currency
        - phone
        - operator
    TransferBody:
      type: object
      properties:
        amount:
          type: number
          minimum: 100
        fromCurrency:
          type: string
          enum: [XAF, XOF, CDF]
        toCurrency:
          type: string
          enum: [XAF, XOF, CDF]
      required:
        - amount
        - fromCurrency
        - toCurrency
    FeePreviewBody:
      type: object
      properties:
        amount:
          type: number
        country:
          type: string
          enum: [CM, SN, CD]
        operator:
          type: string
          enum: [MTN, ORANGE, MOOV, WAVE]
        type:
          type: string
          enum: [DEPOSIT, WITHDRAWAL, TRANSFER]
      required:
        - amount
        - country
        - operator
        - type
    FeeBreakdown:
      type: object
      properties:
        grossAmount:
          type: number
        feeRate:
          type: number
        feeAmount:
          type: number
        netAmount:
          type: number
        currency:
          type: string
        operator:
          type: string
        country:
          type: string
      required:
        - grossAmount
        - feeRate
        - feeAmount
        - netAmount
        - currency
        - operator
        - country
    WebhookBody:
      type: object
      properties:
        reference:
          type: string
        status:
          type: string
          enum: [SUCCESS, FAILED]
        providerReference:
          type: string
        metadata:
          type: object
          nullable: true
          additionalProperties: true
      required:
        - reference
        - status
...