# lib/api-zod/src/generated/types/transferBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YookPay African Payment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TransferBodyFromCurrency } from "./transferBodyFromCurrency";
import type { TransferBodyToCurrency } from "./transferBodyToCurrency";

export interface TransferBody {
  /** @minimum 100 */
  amount: number;
  fromCurrency: TransferBodyFromCurrency;
  toCurrency: TransferBodyToCurrency;
}