# lib/api-zod/src/generated/types/webhookBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YookPay African Payment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { WebhookBodyMetadata } from "./webhookBodyMetadata";
import type { WebhookBodyStatus } from "./webhookBodyStatus";

export interface WebhookBody {
  reference: string;
  status: WebhookBodyStatus;
  providerReference?: string;
  metadata?: WebhookBodyMetadata;
}