# lib/api-zod/src/generated/types/withdrawBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YookPay African Payment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { WithdrawBodyCurrency } from "./withdrawBodyCurrency";
import type { WithdrawBodyOperator } from "./withdrawBodyOperator";

export interface WithdrawBody {
  /** @minimum 100 */
  amount: number;
  currency: WithdrawBodyCurrency;
  phone: string;
  operator: WithdrawBodyOperator;
}